Refactor - SkywardEyeProphetsEffect is shared so moved to new RevealTopLandToBattlefieldElseHandEffect
